requirements.txt
amqp==5.2.0
attrs==23.2.0
asttokens==2.4.1
billiard==4.2.0
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
docopt==0.6.2
executing==2.0.1
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
h11==0.14.0
idna==3.6
ipython==8.12.3
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.3.7
MarkupSafe==2.1.5
outcome==1.3.0.post0
packaging==24.0
parso==0.8.4
pure-eval==0.2.2
python-dotenv==1.0.1
pycparser==2.21
pygame==2.5.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
requests==2.32.3
selenium==4.18.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.30
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.3
wsproto==1.2.0
yarg==0.1.9
APScheduler==3.10.4
requests==2.32.3
